chore(lint): fix lint warnings
