+++ packages/cli/src/dev-portal/Router.tsx
@@ -1,27 +1,6 @@
-import { createBrowserRouter, Outlet, RouterProvider, useParams } from 'react-router-dom';
-import AppLoader from './AppLoader';
-import Header from './Header';
-
-import { PersonProvider } from '@equinor/fusion-react-person';
-import { usePersonResolver } from './usePersonResolver';
-
-const Root = () => {
-    const personResolver = usePersonResolver();
-    return (
-        <div style={{ fontFamily: 'Equinor' }}>
-            <PersonProvider resolve={personResolver}>
-                <Header />
-                <Outlet />
-            </PersonProvider>
-        </div>
-    );
-};
-
-// eslint-disable-next-line react/no-multi-comp
-const AppRoute = () => {
-    const { appKey } = useParams();
-    return appKey ? <AppLoader appKey={appKey} /> : null;
-};
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { AppRoute } from './components/AppRoute';
+import { Root } from './components/Root';
 
 const router = createBrowserRouter([
     {
