+++ packages/cli/src/dev-portal/components/FusionLogo.tsx
