+++ packages/react/legacy-interopt/src/components/LegacyFusionWrapper.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable react/no-multi-comp */
 import {
     PropsWithChildren,
